Update mqtt_client
Much needed retooling of mqtt_client
Completely changed the setup process to a more sane level
It now sets defaults in self.config
If there are command line arguments they replace the default value
if there is not a command line argument and there is a config file value, it replaces the default
otherwise, defaults are used.

This seems to make things much more stable.

You can now use both a JSON file or a YAML file now

mqtt_topics can be an array of strings or a comma dilimited string and it will automatically parse it into an array if its a string.