More logging and Error Catching
add `/` to self.snap and self.clip and replace `//api` with `/api` to make sure the fetch url is correct.
Add more logging to the Fetch functions
Standardize Log Output
Stnadardize returns to a dict with values of error, msg, exception
Change camera and frigate instance name to a input/text box instead of hidden.
Change show to a checkbox instead of actually printing True or False in a textbox